+++ HISTORY.md
@@ -5,6 +5,7 @@
 * Add `autoclose` option to `CacheBackend` to close backend connections when the session context exits.
   * Enabled by default for SQLite backend, and disabled by default for other backends.
 * `python-forge` is no longer required and is now an optional dependency
+* Fix reading response content multiple times for memory backend
 
 ## 0.8.2 (2023-07-14)
 * Add some missing type annotations to backend classes